chore: renamed nwd.ts to fetch to indicate it's function
